+++ frontend/src/features/map_planning/components/timeline/Timeline.tsx
@@ -63,7 +63,10 @@ export function Timeline({ onSelectDate, defaultDate }: TimelineProps) {
   );
 
   return (
-    <form className="flex justify-center gap-2 border-t-2 border-neutral-700 py-1">
+    <form
+      className="flex justify-center gap-2 border-t-2 border-neutral-700 py-1"
+      data-tourid="timeline"
+    >
       <SimpleFormInput
         aria-invalid={submitState === 'error'}
         type="date"
@@ -71,6 +74,7 @@ export function Timeline({ onSelectDate, defaultDate }: TimelineProps) {
         labelText={t('timeline:change_date')}
         register={register}
         title={t('timeline:change_date_hint')}
+        data-tourid="date_picker"
       />
 
       {submitState === 'loading' && (